requirements.txt
# Core ML & Computer Vision
opencv-python>=4.8.0
mediapipe>=0.10.0
scikit-learn>=1.3.0
numpy>=1.24.0

# UI Framework
PySide6>=6.5.0

# Environment config (optional)
python-dotenv>=1.0.0
